simplify tree view setup
