+++ packages/plugin-core/src/views/DendronTreeViewV2.ts
@@ -1,7 +1,6 @@
 import {
   DendronTreeViewKey,
   DMessageEnum,
-  getWebTreeViewEntry,
   NoteProps,
   NoteUtils,
   OnDidChangeActiveTextEditorMsg,
@@ -64,14 +63,12 @@ export class DendronTreeViewV2 implements vscode.WebviewViewProvider {
     const start = process.hrtime();
     Logger.info({ ctx, msg: "enter", start });
 
-    webviewView.webview.options = {
-      enableScripts: true,
-      enableCommandUris: false,
-      localResourceRoots: WebViewUtils.getLocalResourceRoots(
-        getExtension().context
-      ),
-    };
-    webviewView.webview.html = await this._getHtmlForWebview(webviewView);
+    WebViewUtils.prepareTreeView({
+      ext: getExtension(),
+      key: DendronTreeViewKey.TREE_VIEW_V2,
+      webviewView,
+    });
+
     const duration = getDurationMilliseconds(start);
     Logger.info({ ctx, msg: "genHtml:post", duration });
     webviewView.webview.onDidReceiveMessage(async (msg: TreeViewMessage) => {
@@ -142,20 +139,5 @@ export class DendronTreeViewV2 implements vscode.WebviewViewProvider {
         source: "vscode",
       } as OnDidChangeActiveTextEditorMsg);
     }
-  }
-
-  private _getHtmlForWebview(_webview: vscode.WebviewView) {
-    const viewEntry = getWebTreeViewEntry(DendronTreeViewKey.TREE_VIEW_V2);
-    const name = viewEntry.bundleName;
-    const webViewAssets = WebViewUtils.getJsAndCss(name);
-    const ext = getExtension();
-    const port = ext.port!;
-    const html = WebViewUtils.getWebviewContent({
-      ...webViewAssets,
-      port,
-      wsRoot: ext.getEngine().wsRoot,
-      panel: _webview,
-    });
-    return html;
   }
 }
