+++ packages/plugin-core/src/views/utils.ts
@@ -6,11 +6,12 @@ import {
   getStage,
   NoteProps,
   OnDidChangeActiveTextEditorMsg,
+  getWebTreeViewEntry,
 } from "@dendronhq/common-all";
 import _ from "lodash";
 import * as vscode from "vscode";
 import { Logger } from "../logger";
-import { getExtension, getDWorkspace } from "../workspace";
+import { getExtension, getDWorkspace, DendronExtension } from "../workspace";
 import path from "path";
 import { findUpTo, WebViewCommonUtils } from "@dendronhq/common-server";
 import { VSCodeUtils } from "../vsCodeUtils";
@@ -61,7 +62,6 @@ export class WebViewUtils {
    *
    * @param panel: required to convert asset URLs to VSCode Webview Extension format
    * @returns
-   *
    */
   static getWebviewContent({
     jsSrc,
@@ -97,7 +97,38 @@ export class WebViewUtils {
     });
     return out;
   }
+  static prepareTreeView({
+    ext,
+    key,
+    webviewView,
+  }: {
+    ext: DendronExtension;
+    key: DendronTreeViewKey;
+    webviewView: vscode.WebviewView;
+  }) {
+    const viewEntry = getWebTreeViewEntry(key);
+    const name = viewEntry.bundleName;
+    const webViewAssets = WebViewUtils.getJsAndCss(name);
+    const port = ext.port!;
+    webviewView.webview.options = {
+      enableScripts: true,
+      enableCommandUris: false,
+      localResourceRoots: WebViewUtils.getLocalResourceRoots(ext.context),
+    };
+    const html = WebViewUtils.getWebviewContent({
+      ...webViewAssets,
+      port,
+      wsRoot: ext.getEngine().wsRoot,
+      panel: webviewView,
+    });
+    webviewView.webview.html = html;
+  }
 
+  /**
+   * @deprecated Use `{@link WebviewUtils.getWebviewContent}`
+   * @param param0
+   * @returns
+   */
   static genHTMLForView = async ({
     title,
     view,
